add user_cart

+++ userinterface/models.py
@@ -1,4 +1,5 @@
 from django.db import models
+from django.contrib.auth.models import User
 
 
 # Create your models here.
@@ -6,9 +7,11 @@ class product(models.Model):
   product_name = models.CharField(max_length=100)
   description = models.TextField()
   brand = models.CharField(max_length=100)
-  price = models.DecimalField(max_digits=10,decimal_places=2)
+  price = models.IntegerField()
   stock = models.IntegerField()
   picture = models.ImageField(upload_to="product",blank=True)
 
   def __str__(self) -> str:
     return f'{self.product_name}'
+  
+
